BACKEND/requirements.txt
pandas
fastapi
uvicorn
python-dotenv
pydantic
email-validator
pymysql
SQLAlchemy
PyJWT
alembic
fastapi-cors
bcrypt
psycopg2-binary